Add files via upload
In this implementation, we have an example of priority queue (min binary heap). It was made for fun, but it can be used in any practical approach.

Using some IDE, create package "main", insert into it the classes: Main.java and PriorityQueue.java  and have fun! :D

Specifically, the PriorityQueue class has the methods: constructMinHeap(), isEmpty(), decreasekey() and extractMin(). The Main class has just an example of object (PriorityQueue type) creation, initiation and manipulation of some integer samples.